refactor: move shared query-plan logic to fabric_dw.plan neutral module
mcp/tools/sql_exec.py imported parse_showplan, operator_to_dict, and
render_plan_mermaid from fabric_dw.cli.*, making the CLI presentation
layer a hard runtime dependency of the MCP server (inverted layering).

Create fabric_dw/plan/ with three modules:
- _parse.py  - SHOWPLAN XML parser (PlanOperator, parse_showplan, humanise_rows)
- _mermaid.py - Mermaid flowchart renderer (render_plan_mermaid)
- _render.py  - dict serialiser (operator_to_dict)

Update mcp/tools/sql_exec.py to import exclusively from fabric_dw.plan.*.
The plan/ package carries no cli or mcp imports, breaking the cycle.

Turn cli/_plan_parse.py and cli/_plan_mermaid.py into compatibility shims
that re-export the full symbol set (including private helpers used by unit
tests) so existing CLI commands and all test files continue to work
without modification.

cli/_plan_render.py retains render_plan_tree (Rich terminal output, CLI-only)
and re-exports operator_to_dict from fabric_dw.plan._render for backward
compatibility. cli/_plan_dot.py and cli/_plan_svg.py are updated to import
PlanOperator/humanise_rows directly from fabric_dw.plan._parse.
